pyproject.toml
[tool.poetry]
name = "kkp-backend"
version = "0.1.0"
description = ""
authors = ["RuslanUC <dev_ruslan_uc@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
pydantic = {extras = ["email"], version = "^2.11.4"}
tortoise-orm = {extras = ["accel", "asyncmy"], version = "^0.25.0"}
s3lite = "^0.1.8"
pydantic-settings = "^2.9.1"
aerich = "^0.8.2"
bcrypt = "^4.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
